backend/requirements.txt
flask==3.0.2
numba
flask_cors
mysql-connector-python==8.3.0
uwsgi==2.0.24